+++ frontend/lib/main.dart
@@ -5,9 +5,8 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:orot/firebase_options.dart';
-import 'package:orot/pages/coordinator/volunteersList/volunteers_page.dart';
 import 'package:orot/pages/login/login_page.dart';
-import 'package:orot/user_provider.dart';
+import 'package:orot/providers/user_provider.dart';
 import 'package:provider/provider.dart';
 
 Future<void> main() async {
@@ -46,7 +45,7 @@ class OrotApp extends StatelessWidget {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       theme: ThemeData(fontFamily: 'Open Sans'),
-      home: user == null ? const LoginPage() : const VolunteersPage(),
+      home: user == null ? const LoginPage() : const LoginPage(),
     );
   }
 }
