+++ frontend/lib/pages/admin/admin_page.dart
@@ -6,8 +6,8 @@ import 'package:orot/pages/admin/add_coordinator_page.dart';
 import 'package:orot/pages/admin/add_district_page.dart';
 import 'package:orot/pages/admin/add_family_page.dart';
 import 'package:orot/pages/admin/add_volunteer_page.dart';
-import 'package:orot/pages/home/home_page.dart';
-import 'package:orot/user_provider.dart';
+import 'package:orot/pages/volunteer/home/home_page.dart';
+import 'package:orot/providers/user_provider.dart';
 import 'package:provider/provider.dart';
 
 class AdminPage extends StatelessWidget {
