+++ frontend/lib/pages/admin/components/districts_dropdown.dart
@@ -6,6 +6,7 @@ class DistrictsDropdown extends StatefulWidget {
   ValueChanged<String?> onSelectedIdChange;
   List<District> districts;
   String selectedDistrictId;
+
   DistrictsDropdown(
       {super.key,
       required this.districts,
