+++ frontend/lib/pages/volunteer/home/home_page.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:orot/components/main_button.dart';
 import 'package:orot/components/visit_card.dart';
 import 'package:orot/pages/admin/admin_page.dart';
-import 'package:orot/pages/home/home_label.dart';
-import 'package:orot/pages/home/home_title.dart';
-import 'package:orot/pages/home/visits_list.dart';
-import 'package:orot/user_provider.dart';
+import 'package:orot/pages/volunteer/home/home_label.dart';
+import 'package:orot/pages/volunteer/home/home_title.dart';
+import 'package:orot/pages/volunteer/home/visits_list.dart';
+import 'package:orot/providers/user_provider.dart';
 import 'package:provider/provider.dart';
 
 class HomePage extends StatefulWidget {
