+++ frontend/lib/pages/volunteer/new_visit/new_visit_page.dart
@@ -5,8 +5,8 @@ import 'package:intl/intl.dart';
 import 'package:orot/components/main_button.dart';
 import 'package:orot/modal/calendar/event.dart';
 import 'package:orot/modal/calendar/event_time.dart';
-import 'package:orot/pages/home/home_page.dart';
-import 'package:orot/pages/newVisit/field.dart';
+import 'package:orot/pages/volunteer/home/home_page.dart';
+import 'package:orot/pages/volunteer/new_visit/field.dart';
 import 'package:orot/services/calender_service.dart';
 import 'package:orot/services/google_auth.dart';
 import 'package:orot/services/volunteer_service.dart';
