+++ frontend/lib/pages/volunteer/visits_history/visits_history_app_bar.dart
