+++ frontend/lib/services/auth_service.dart
@@ -3,8 +3,8 @@ import 'package:cloud_functions/cloud_functions.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:orot/pages/home/home_page.dart';
 import 'package:orot/pages/login/login_page.dart';
+import 'package:orot/pages/volunteer/home/home_page.dart';
 
 class AuthService {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
